+++ src/PhpPresentation/Writer/PowerPoint2007/PptCharts.php
@@ -2014,8 +2014,14 @@ class PptCharts extends AbstractDecoratorWriter
         $objWriter->writeAttribute('val', 'nextTo');
         $objWriter->endElement();
 
-        // c:txPr
-        $objWriter->startElement('c:txPr');
+        // c:title
+        $objWriter->startElement('c:title');
+
+        // c:tx
+        $objWriter->startElement('c:tx');
+
+        // c:rich
+        $objWriter->startElement('c:rich');
 
         // a:bodyPr
         $objWriter->writeElement('a:bodyPr', null);
@@ -2081,7 +2087,13 @@ class PptCharts extends AbstractDecoratorWriter
         // ## a:p
         $objWriter->endElement();
 
-        // ## c:txPr
+        // ## c:rich
+        $objWriter->endElement();
+
+        // ## c:tx
+        $objWriter->endElement();
+
+        // ## c:title
         $objWriter->endElement();
 
         // c:crossAx