+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - PowerPoint2007 Writer : Fixed the format value in Pie Chart - @Napryc GH-212
 - PowerPoint2007 Writer : The presentation need repairs on Mac @jrking4 GH-266 GH-276
 - PowerPoint2007 Writer : Fix for PowerPoint2007 Writer (Need repair) @Progi1984 GH-266 GH-274 GH-276 GH-282 GH-302
+- PowerPoint2007 Writer : Fixed the axis title in bar chart - @pgee70 GH-267
 
 ## 0.7.0 - 2016-09-12